+++ api/graphql/makeModels.js
@@ -113,6 +113,7 @@ export default async function makeModels (userId, isAdmin) {
         'moderatedCommunityMemberships',
         'locationObject',
         {posts: {querySet: true}},
+        {projects: {querySet: true}},
         {comments: {querySet: true}},
         {skills: {querySet: true}},
         {skillsToLearn: {querySet: true}},